Extracted serde code for Split into separate file Added serde code for Area
